Creates api module, and moves statistics-xxx modules into statistics folder
