+++ statistics/statistics-api/build.gradle
@@ -1,0 +1,17 @@
+plugins {
+    id 'java-library'
+    id 'kotlin'
+}
+
+apply from: "$rootProject.projectDir/spotless.gradle"
+
+java {
+    sourceCompatibility = JavaVersion.VERSION_1_8
+    targetCompatibility = JavaVersion.VERSION_1_8
+}
+
+dependencies {
+    implementation Google.dagger
+    implementation Kotlin.stdlib.jdk7
+}
+
