+++ statistics/statistics-impl/build.gradle
@@ -29,6 +29,7 @@ dependencies {
     implementation project(path: ':common')
     implementation project(path: ':di')
     implementation project(path: ':app-build-config-api')
+    implementation project(path: ':statistics-api')
 
     implementation Kotlin.stdlib.jdk7
     implementation KotlinX.coroutines.core
