+++ statistics/statistics-impl/src/main/AndroidManifest.xml
