+++ statistics/statistics-impl/src/main/java/com/duckduckgo/app/statistics/VariantManager.kt
