+++ statistics/statistics-impl/src/main/java/com/duckduckgo/app/statistics/pixels/ExceptionPixel.kt
